pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["jacksonsr451 <jacksonsr451@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
colorlog = "^6.9.0"
python-json-logger = "^2.0.7"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
httpx = "^0.27.2"
redis = "^5.2.0"
pydantic = "^2.10.2"
asyncpg = "^0.30.0"
pyjwt = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
